+++ prisma/migrations/20220910135140_removed_currency_and_view_mode_from_user/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "Settings" DROP COLUMN "currency",
+DROP COLUMN "viewMode";